chore: move use-tab-nav-touch to composables
